Abandon the concept of waiting in sleep mode for a frame to arrive, which would be announced by an interrupt from the enc28j60. But this mechanism is so fragile, even with some workarounds, it can't be made reliable.
So removed the sleep mode when waiting for a frame and return
immediately if there are no frames in the buffer. Then, in
main issue a sleep mode, which is then waken up from by any
interrupt (either periodic timer1 122 Hz or watchdog or frame
arrival from enc28j60) and frame arrival is checked.

This means most of the time after wake up there will be no
new frames, but at least it allows the cpu to go into  sleep
mode ~120 times per second instead of just running round in
busy mode.

The main advantage of this approach is that a newly arrived
frame is detected quickly, even when the interrupt from the
enc28j60 fails.